Synchronization initialization.
The initialization step of the synchronization loop has been
implemented. It precomputes the slides' keypoints and descriptors
as well as matching the first frame with the first slide to find
the presentation projection in the footage, (which will be used
to speed up the tracking and matching of the presentation in the
next frames).

The types of some variables have been modified to align them with
related operations (so the amount of conversions are minimized,
improving performance and quality), e.g. int to double,
double to float.

The SyncLoop now has a feature detector and a matcher, as well as
cached keypoints and descriptors to avoid computing them twice.

BRISK has been chosen as a feature detector. In preliminary
experiments it shows superior performance with text, which are
common on slides. The matcher is based on the Hamming distance
and brute force, filtered by RANSAC and a maximum distance ratio
between the best match and the second best match.

+++ src/IMhelpers.cpp
@@ -98,15 +98,15 @@ std::vector<Magick::Image*> readpdf_im(string filename, int framewidth, int fram
 		return slides;
 	}
 	
-	double resolution = 0;
+	float resolution = 0;
 	
 	// fit the slides bounding box to the frame
-	if (((double) framewidth) / testwidth < ((double) frameheight) / testheight) {
+	if (((float) framewidth) / testwidth < ((float) frameheight) / testheight) {
 		// pagewidth = width / resolution is constant and then solve for resolution
-		resolution = ((double) testresolution) * framewidth / testwidth;
+		resolution = ((float) testresolution) * framewidth / testwidth;
 	}
 	else {
-		resolution = ((double) testresolution) * frameheight / testheight;
+		resolution = ((float) testresolution) * frameheight / testheight;
 	}
 	
 	const int antialias = 4;
